feat: Pixiu Metrics Implementation #433
