+++ pkg/metrics/collector/api_health_test.go
@@ -1,44 +1,203 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package collector
 
 import (
-	"bytes"
+	"encoding/json"
 	"fmt"
+	"io/ioutil"
 	"log"
 	"net/http"
-	"net/http/httptest"
 	"net/url"
-	"testing"
+	"path"
+)
+import (
+	"github.com/prometheus/client_golang/prometheus"
 )
 
-func TestApiHealth(t *testing.T) {
+import (
+	"github.com/apache/dubbo-go-pixiu/pkg/metrics/global"
+)
 
-	tcs := map[string]string{
-		"1.7.6": ` {"share":{"total":1000,"successful":998,"failed":2},"api_stats":[{"api_name":"Api_1","api_requests":10,"api_requests_latency":1},{"api_name":"Api_12","api_requests":20,"api_requests_latency":2}]}`,
-		"2.4.5": ` {"share":{"total":2000,"successful":1998,"failed":2},"api_stats":[{"api_name":"Api_1","api_requests":20,"api_requests_latency":2},{"api_name":"Api_12","api_requests":20,"api_requests_latency":3}]}`,
+// A collection of many Api states
+type ApiStatsResponse struct {
+	Share    ApiStatShare `json:"share"`
+	ApiStats []ApiStat    `json:"api_stats"`
+}
+
+//The state shared by the Api
+type ApiStatShare struct {
+	Total      int64 `json:"total"`
+	Successful int64 `json:"successful"`
+	Failed     int64 `json:"failed"`
+}
+
+// Status of a single Api
+type ApiStat struct {
+	ApiName            string `json:"api_name"`
+	ApiRequests        int64  `json:"api_requests"`
+	ApiRequestsLatency int64  `json:"api_requests_latency"`
+}
+
+// Implement the structure for collecting Api indicators
+type apiMetric struct {
+	Type   prometheus.ValueType
+	Desc   *prometheus.Desc
+	Value  func(dataStreamStats ApiStat) float64
+	Labels func(dataStreamStats ApiStat) []string
+}
+
+var (
+	defaultApiMetricLabels      = []string{"api"}
+	defaultApiMetricLabelValues = func(apiStat ApiStat) []string {
+		return []string{apiStat.ApiName}
 	}
-	for ver, out := range tcs {
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			fmt.Fprintln(w, out)
-		}))
-		defer ts.Close()
+)
 
-		u, err := url.Parse(ts.URL)
+// ApiHealth Struct
+type ApiHealth struct {
+	logger log.Logger
+	client *http.Client
+	url    *url.URL
+
+	up                prometheus.Gauge
+	totalScrapes      prometheus.Counter
+	jsonParseFailures prometheus.Counter
+
+	apiMetrics []*apiMetric
+}
+
+// NewApiHealth defines  Api Health Prometheus metrics
+func NewApiHealth(logger log.Logger, client *http.Client, url *url.URL) *ApiHealth {
+	return &ApiHealth{
+		logger: logger,
+		client: client,
+		url:    url,
+
+		up: prometheus.NewGauge(prometheus.GaugeOpts{
+			Name: prometheus.BuildFQName(global.Namespace, "api_stats", "up"),
+			Help: "Was the last scrape of the Pixiu Api Stat Data successful.",
+		}),
+		totalScrapes: prometheus.NewCounter(prometheus.CounterOpts{
+			Name: prometheus.BuildFQName(global.Namespace, "api_stats", "total_scrapes"),
+			Help: "Current total Pixiu Api scrapes.",
+		}),
+		jsonParseFailures: prometheus.NewCounter(prometheus.CounterOpts{
+			Name: prometheus.BuildFQName(global.Namespace, "api_stats", "json_parse_failures"),
+			Help: "Number of errors while parsing JSON.",
+		}),
+		apiMetrics: []*apiMetric{
+			{
+				Type: prometheus.CounterValue,
+				Desc: prometheus.NewDesc(
+					prometheus.BuildFQName(global.Namespace, "api_stats", "api_Requests_total"),
+					"Number of Api Requests",
+					defaultApiMetricLabels, nil,
+				),
+				Value: func(apiStats ApiStat) float64 {
+					return float64(apiStats.ApiRequests)
+				},
+				Labels: defaultApiMetricLabelValues,
+			},
+		},
+	}
+}
+
+// Describe adds api metrics descriptions
+func (ds *ApiHealth) Describe(ch chan<- *prometheus.Desc) {
+	for _, metric := range ds.apiMetrics {
+		ch <- metric.Desc
+	}
+
+	ch <- ds.up.Desc()
+	ch <- ds.totalScrapes.Desc()
+	ch <- ds.jsonParseFailures.Desc()
+}
+
+func (ds *ApiHealth) fetchAndDecodeApiStats() (ApiStatsResponse, error) {
+	var dsr ApiStatsResponse
+
+	u := *ds.url
+	u.Path = path.Join(u.Path, "/_api_health/*/_stats")
+	res, err := ds.client.Get(u.String())
+	if err != nil {
+		return dsr, fmt.Errorf("failed to get api stats health from %s://%s:%s%s: %s",
+			u.Scheme, u.Hostname(), u.Port(), u.Path, err)
+	}
+
+	defer func() {
+		err = res.Body.Close()
 		if err != nil {
-			t.Fatalf("Failed to parse URL: %s", err)
+			_ = ds.logger.Output(
+				2,
+				"msg:"+"failed to close http.Client"+"err:"+err.Error(),
+			)
 		}
+	}()
 
-		buf := &bytes.Buffer{}
-		logger := log.New(buf, "", log.Lshortfile|log.LstdFlags)
+	if res.StatusCode != http.StatusOK {
+		return dsr, fmt.Errorf("HTTP Request failed with code %d", res.StatusCode)
+	}
 
-		c := NewApiHealth(*logger, http.DefaultClient, u)
-		chr, err := c.fetchAndDecodeApiStats()
-		if err != nil {
-			t.Fatalf("Failed to fetch or decode api health: %s", err)
+	bts, err := ioutil.ReadAll(res.Body)
+	if err != nil {
+		ds.jsonParseFailures.Inc()
+		return dsr, err
+	}
+
+	if err := json.Unmarshal(bts, &dsr); err != nil {
+		ds.jsonParseFailures.Inc()
+		return dsr, err
+	}
+
+	return dsr, nil
+}
+
+// Collect gets Api metric values
+func (ds *ApiHealth) Collect(ch chan<- prometheus.Metric) {
+	ds.totalScrapes.Inc()
+	defer func() {
+		ch <- ds.up
+		ch <- ds.totalScrapes
+		ch <- ds.jsonParseFailures
+	}()
+
+	dataStreamStatsResp, err := ds.fetchAndDecodeApiStats()
+	if err != nil {
+		ds.up.Set(0)
+		_ = ds.logger.Output(
+			2,
+			"msg"+"failed to fetch and decode data stream stats"+"err:"+err.Error(),
+		)
+		return
+	}
+
+	ds.up.Set(1)
+
+	for _, metric := range ds.apiMetrics {
+		for _, apistat := range dataStreamStatsResp.ApiStats {
+			fmt.Printf("Metric: %+v", apistat)
+			ch <- prometheus.MustNewConstMetric(
+				metric.Desc,
+				metric.Type,
+				metric.Value(apistat),
+				metric.Labels(apistat)...,
+			)
 		}
-		t.Logf("[%s] Api Health Response: %+v", ver, chr)
-		if chr.ApiStats[0].ApiName == "Api_1" {
-			fmt.Println("ok")
-		}
-
 	}
 }
